Working on the DeviceSet
Now the list of devices is of type DevList and that is parametrized: I can still use an array of Abstract elements but I can use anything that has `length` and can be indexed. Tuples are possible. Might help in case of the need for performance.

I am also working on measdata.jl. That is a structure that stores data acquired. I am not shure I will use this but it has a few ideas for the future. I might simply remove it.